# Document/swagger-insure.v1.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "RFQ"
    },
    "servers": [
        {
            "url": "http://192.168.3.56:9763",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/rfq/clientList/updateClient/{id}": {
            "put": {
                "tags": [
                    "client-list-controller"
                ],
                "operationId": "updateClient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientListChildDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClientListChildDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/rfq/clientList/add": {
            "post": {
                "tags": [
                    "client-list-controller"
                ],
                "operationId": "createClient",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientListDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClientListDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/rfq/getAgebandingReport/{id}": {
            "get": {
                "tags": [
                    "age-binding-report-controller"
                ],
                "operationId": "generatePdf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "format": "byte"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/rfq/clientList/getClient/{id}": {
            "get": {
                "tags": [
                    "client-list-controller"
                ],
                "operationId": "getClientDetailsById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClientListDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/templates/{id}": {
            "get": {
                "tags": [
                    "template-controller"
                ],
                "operationId": "getFileDataById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/rfq/clientList/deleteClientByClientListId/{id}": {
            "delete": {
                "tags": [
                    "client-list-controller"
                ],
                "operationId": "deleteClientByClientListId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClientListChildDto": {
                "required": [
                    "clientName",
                    "location"
                ],
                "type": "object",
                "properties": {
                    "clientName": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    }
                }
            },
...
            "ClientListDto": {
                "required": [
                    "clientName",
                    "insuranceCompanyId",
                    "locationId",
                    "productId"
                ],
                "type": "object",
                "properties": {
                    "clientName": {
                        "type": "string"
                    },
                    "locationId": {
                        "type": "string"
                    },
                    "productId": {
                        "type": "string"
                    },
                    "insuranceCompanyId": {
                        "type": "string"
                    },
                    "policyType": {
                        "type": "string"
                    },
                    "tpaId": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}